fixed RaitingInfo property type

+++ src/TvDbSharper/RestClient/Models/RatingsInfo.cs
@@ -2,7 +2,7 @@ namespace TvDbSharper.RestClient.Models
 {
     public class RatingsInfo
     {
-        public int Average { get; set; }
+        public decimal Average { get; set; }
 
         public int Count { get; set; }
     }
